# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Veloce Wear SCM Simulation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  studentId: string;
  password: string;
  confirmPassword: string;
  section?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  student: "student",
  instructor: "instructor",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  studentId: string;
  section?: string | null;
  role: UserProfileRole;
}

export interface AuthResponse {
  user: UserProfile;
  message: string;
}

export type ModuleInfoStatus =
  (typeof ModuleInfoStatus)[keyof typeof ModuleInfoStatus];

export const ModuleInfoStatus = {
  Locked: "Locked",
  Not_started: "Not started",
  In_progress: "In progress",
  Submitted: "Submitted",
} as const;

export type ModuleInfoWindow =
  (typeof ModuleInfoWindow)[keyof typeof ModuleInfoWindow];

export const ModuleInfoWindow = {
  Open: "Open",
  Closed: "Closed",
} as const;

export interface ModuleInfo {
  key: string;
  title: string;
  maxScore: number;
  status: ModuleInfoStatus;
  window: ModuleInfoWindow;
  isUnlocked: boolean;
  score?: number | null;
  submittedAt?: string | null;
  practiceRunCount: number;
}

export interface DashboardData {
  modules: ModuleInfo[];
  totalScore: number;
  maxTotal: number;
  totalPracticeRuns: number;
  userName: string;
}

export interface SimulationRun {
  runNumber: number;
  score: number;
  isFinal: boolean;
  createdAt: string;
}

export type ModuleDataFinalSubmission = {
  score?: number;
  maxScore?: number;
  submittedAt?: string;
} | null;

export interface ModuleData {
  moduleKey: string;
  isSubmitted: boolean;
  isUnlocked: boolean;
  finalSubmission?: ModuleDataFinalSubmission;
  recentRuns: SimulationRun[];
  practiceCount: number;
}

export interface RunResult {
  runNumber: number;
  score: number;
  isFinal: boolean;
  message: string;
}

export interface StudentGradeRow {
  id: number;
  name: string;
  email: string;
  studentId: string;
  section: string;
  m1Score: number;
  m1Submitted?: string | null;
  m2Score: number;
  m2Submitted?: string | null;
  m3Score: number;
  m3Submitted?: string | null;
  total: number;
}

export interface GradebookData {
  students: StudentGradeRow[];
  sections: string[];
  totalStudents: number;
}

export interface ModuleWindow {
  moduleKey: string;
  title: string;
  startAt: string;
  endAt: string;
  isEnabled: boolean;
}

export interface ExtensionRow {
  id: number;
  studentName: string;
  studentId: string;
  moduleKey: string;
  extendedEnd: string;
  note: string;
}

export interface SettingsData {
  windows: ModuleWindow[];
  extensions: ExtensionRow[];
}

export type UpdateWindowsRequestWindowsItem = {
  moduleKey: string;
  startAt: string;
  endAt: string;
  isEnabled: boolean;
};

export interface UpdateWindowsRequest {
  windows: UpdateWindowsRequestWindowsItem[];
}

export interface AddExtensionRequest {
  studentId: string;
  moduleKey: string;
  extendedEnd: string;
  note?: string;
}

export type GetGradebookParams = {
  search?: string;
  section?: string;
};