# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Veloce Wear SCM Simulation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /instructor/gradebook:
    get:
      operationId: getGradebook
      tags: [instructor]
      summary: Get gradebook with all students
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: section
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Gradebook data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradebookData"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /instructor/settings/windows:
    put:
      operationId: updateModuleWindows
      tags: [instructor]
      summary: Update module time windows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWindowsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /instructor/extensions:
    post:
      operationId: addExtension
      tags: [instructor]
      summary: Add per-student extension
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddExtensionRequest"
      responses:
        "200":
          description: Extension added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /instructor/extensions/{extensionId}:
    delete:
      operationId: removeExtension
      tags: [instructor]
      summary: Remove a student extension
      parameters:
        - name: extensionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extension removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        studentId:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        section:
          type: string
      required: [name, email, studentId, password, confirmPassword]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        message:
          type: string
      required: [user, message]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        studentId:
          type: string
        section:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, instructor]
      required: [id, name, email, studentId, role]
...
    StudentGradeRow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        studentId:
          type: string
        section:
          type: string
        m1Score:
          type: integer
        m1Submitted:
          type: string
          nullable: true
        m2Score:
          type: integer
        m2Submitted:
          type: string
          nullable: true
        m3Score:
          type: integer
        m3Submitted:
          type: string
          nullable: true
        total:
          type: integer
      required: [id, name, email, studentId, section, m1Score, m2Score, m3Score, total]
    GradebookData:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/StudentGradeRow"
        sections:
          type: array
          items:
            type: string
        totalStudents:
          type: integer
      required: [students, sections, totalStudents]
...
    UpdateWindowsRequest:
      type: object
      properties:
        windows:
          type: array
          items:
            type: object
            properties:
              moduleKey:
                type: string
              startAt:
                type: string
              endAt:
                type: string
              isEnabled:
                type: boolean
            required: [moduleKey, startAt, endAt, isEnabled]
      required: [windows]
    AddExtensionRequest:
      type: object
      properties:
        studentId:
          type: string
        moduleKey:
          type: string
        extendedEnd:
          type: string
        note:
          type: string
      required: [studentId, moduleKey, extendedEnd]