# lib/api-zod/src/generated/types/studentGradeRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Veloce Wear SCM Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentGradeRow {
  id: number;
  name: string;
  email: string;
  studentId: string;
  section: string;
  m1Score: number;
  m1Submitted?: string | null;
  m2Score: number;
  m2Submitted?: string | null;
  m3Score: number;
  m3Submitted?: string | null;
  total: number;
}