# lib/api-zod/src/generated/types/updateWindowsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Veloce Wear SCM Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateWindowsRequestWindowsItem } from "./updateWindowsRequestWindowsItem";

export interface UpdateWindowsRequest {
  windows: UpdateWindowsRequestWindowsItem[];
}